+++ data.tf
@@ -1,6 +1,6 @@
 locals {
-  kv_db_name = var.kv_db_name
-  kv_db_rg   = var.kv_db_rg
+  kv_name = var.kv_name
+  kv_rg   = var.kv_rg
 
   kv_workflow_name = var.kv_workflow_name
   kv_workflow_rg   = var.kv_workflow_rg
@@ -9,10 +9,10 @@ locals {
 data "azurerm_client_config" "current" {}
 
 data "azurerm_key_vault" "db" {
-  count = var.kv_db_enable ? 1 : 0
+  count = var.kv_create ? 0 : 1
 
-  name                = local.kv_db_name
-  resource_group_name = local.kv_db_rg
+  name                = local.kv_name
+  resource_group_name = local.kv_rg
 }
 
 data "azurerm_key_vault" "sqlhstkv" {
@@ -44,14 +44,14 @@ data "azurerm_storage_account" "saloggingname" {
 }
 
 data "azurerm_virtual_network" "mysql" {
-  count = var.vnet_enable ? 0 : 1
+  count = var.vnet_create ? 0 : 1
 
   name                = var.vnet_name
   resource_group_name = var.vnet_rg
 }
 
 data "azurerm_subnet" "mysql" {
-  count = var.subnet_enable ? 0 : 1
+  count = var.subnet_create ? 0 : 1
 
   name                 = var.subnet_name
   virtual_network_name = var.vnet_name