+++ examples/standard.tf
@@ -3,7 +3,7 @@ provider "azurerm" {
 }
 
 module "mysql_example" {
-  source = "git::https://github.com/canada-ca-terraform-modules/terraform-azurerm-mysql.git?ref=v2.2.0"
+  source = "git::https://github.com/canada-ca-terraform-modules/terraform-azurerm-mysql.git?ref=v3.0.0"
 
   name = "mysqlservername"
   databases = {
@@ -14,16 +14,26 @@ module "mysql_example" {
   }
 
   administrator_login          = "mysqladmin"
-  administrator_login_password = var.administrator_login_password
+  administrator_login_password = "mysql1313"
 
   # active_directory_administrator_object_id = "XX-XXXX-XXXX-XXX-XXX"
   # active_directory_administrator_tenant_id = "XX-XXXX-XXXX-XXX-XXX"
 
-  kv_db_enable    = false
-  # kv_db_tenant_id = "XX-XXXX-XXXX-XXX-XXX"
-  # kv_db_name             = "XXXXX"
-  # kv_db_rg               = "XX-XXXX-XXXX-XXX-XXX"
+  #########################################################
+  # kv_create
+  # => `true` then enable creation of new key vault
+  # => `false` then point to existing key vault
+  #########################################################
+  kv_create       = true
+  kv_tenant_id = "XX-XXXX-XXXX-XXX-XXX"
+  kv_name      = "mysql-keyvault"
+  kv_rg        = "XX-XXXX-XXXX-XXX-XXX"
 
+  #########################################################
+  # kv_workflow_enable
+  # => ``true` then enable storing pointers to secrets in key vault
+  # => ``false` then store as default
+  #########################################################
   kv_workflow_enable = false
   # kv_workflow_name             = "XXXXX"
   # kv_workflow_rg               = "XX-XXXX-XXXX-XXX-XXX"
@@ -40,25 +50,36 @@ module "mysql_example" {
   ip_rules       = []
   firewall_rules = []
 
-  vnet_cidr   = "172.15.0.0/16"
-  vnet_enable = false
+  #########################################################
+  # vnet_create
+  # => ``true` then enable creation of new vnet
+  # => ``false` then point to existing vnet
+  #########################################################
+  vnet_create = false
+  # vnet_cidr   = "172.15.0.0/16"
   vnet_name   = "mysql-vnet"
   vnet_rg     = "XX-XXXX-XXXX-XXX-XXX"
 
-  subnet_enable           = false
-  subnet_name             = "mysql-subnet"
-  subnet_address_prefixes = ["172.15.8.0/22"]
+  #########################################################
+  # subnet_create
+  # => ``true` then enable creation of new subnet
+  # => ``false` then point to existing subnet
+  #########################################################
+  subnet_create = false
+  subnet_name   = "mysql-subnet"
+  # subnet_address_prefixes = ["172.15.8.0/22"]
 
   public_network_access_enabled    = true
   ssl_enforcement_enabled          = true
   ssl_minimal_tls_version_enforced = "TLS1_2"
 
-  diagnostics = {
-    destination   = ""
-    eventhub_name = ""
-    logs          = ["all"]
-    metrics       = ["all"]
-  }
+  # Needs to be disabled until the following issue is resolved: https://github.com/MicrosoftDocs/azure-docs/issues/32068
+  # diagnostics = {
+  #   destination   = ""
+  #   eventhub_name = ""
+  #   logs          = ["all"]
+  #   metrics       = ["all"]
+  # }
 
   emails = []