+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v3.0.0
+
+- More reflective variable naming
+- Better inline documentation
+
+## v2.2.0
+
+- Allow for key vault creation inside the module
+
 ## v2.1.0
 
 - Added bring your own Virtual Network and Subnet logic
@@ -5,6 +14,6 @@
 - Fixed the logging behavior so properly set to NULL when not enabled
 - Add Networking ACL for the storage accounts
 
-Upgrade path:
+## Upgrade path
 
 > Note: See [MIGRATION.md](MIGRATION.md) if you are updating from previous versions.